code/aleksandr-k3s-istio1/requirements.txt
Flask==2.2.5
gunicorn
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
otlp-stdout-span-exporter

